devtools/client/shared/components/reps/test/mochitest/test_reps_window.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test window rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep tests - window</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");

    const {
      REPS,
      MODE,
      getRep,
    } = browserRequire("devtools/client/shared/components/reps/reps");
    let { Rep, Window } = REPS;

    let gripStub = {
      "type": "object",
      "class": "Window",
      "actor": "server1.conn3.obj198",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 887,
      "preview": {
        "kind": "ObjectWithURL",
        "url": "about:newtab"
      }
    };

    // Test that correct rep is chosen
    is(getRep(gripStub), Window.rep, "Rep correctly selects Window Rep");

    // Test rendering
    const renderedComponent = renderComponent(Rep, { object: gripStub });
    ok(renderedComponent.className.includes("objectBox-Window"), "Window rep has expected class name");
    is(renderedComponent.textContent, "Window about:newtab", "Window rep has expected text content");
    const innerNode = renderedComponent.querySelector(".objectPropValue");
    is(innerNode.textContent, "about:newtab", "Window rep has expected inner HTML structure and text content");

    const tinyRenderedComponent = renderComponent(Window.rep, {
      object: gripStub,
      mode: MODE.TINY,
    });
    is(tinyRenderedComponent.textContent, "Window",
       "Window rep has expected text content in TINY mode");

    const longRenderedComponent = renderComponent(Window.rep, {
      object: gripStub,
      mode: MODE.LONG,
    });
    is(longRenderedComponent.textContent, "Window about:newtab",
       "Window rep has expected text content in LONG mode");

    const displayClassRenderedComponent = renderComponent(Window.rep, {
      object: Object.assign({}, gripStub, {displayClass : "Custom"}),
      mode: MODE.TINY,
    });
    is(displayClassRenderedComponent.textContent, "Custom",
       "Window rep has expected text content in TINY mode with Custom display class");

    const displayClassLongRenderedComponent = renderComponent(Window.rep, {
      object: Object.assign({}, gripStub, {displayClass : "Custom"}),
      mode: MODE.LONG,
      title: "Custom"
    });
    is(displayClassLongRenderedComponent.textContent, "Custom about:newtab",
       "Window rep has expected text content in LONG mode with Custom display class");

    const objectLinkTinyRenderedComponent = renderComponent(Window.rep, {
      object: gripStub,
      mode: MODE.TINY,
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(
      objectLinkTinyRenderedComponent.textContent,
      "*Window*",
      "Window rep has expected text content in TINY mode when an objectLink is passed as a prop"
    );

    const objectLinkLongRenderedComponent = renderComponent(Window.rep, {
      object: gripStub,
      mode: MODE.LONG,
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(
      objectLinkLongRenderedComponent.textContent,
      "*Window* about:newtab",
      "Window rep has expected text content in LONG mode when an objectLink is passed as a prop"
   );
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
